// apps/mobile/src/api/generated/models/modelReasoningCapabilityDto.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * 办伴 Banban API
 * 办伴 Banban 后端接口文档
 * OpenAPI spec version: 1.0
 */
import type { ModelReasoningCapabilityDtoTemperaturePolicy } from './modelReasoningCapabilityDtoTemperaturePolicy';
import type { ModelReasoningCapabilityDtoTopPPolicy } from './modelReasoningCapabilityDtoTopPPolicy';
import type { ReasoningActivationCapabilityDto } from './reasoningActivationCapabilityDto';
import type { ReasoningBudgetCapabilityDto } from './reasoningBudgetCapabilityDto';
import type { ReasoningEffortCapabilityDto } from './reasoningEffortCapabilityDto';
import type { ReasoningSelectionDto } from './reasoningSelectionDto';

export interface ModelReasoningCapabilityDto {
  activation?: ReasoningActivationCapabilityDto;
  effort?: ReasoningEffortCapabilityDto;
  budget?: ReasoningBudgetCapabilityDto;
  defaultSelection?: ReasoningSelectionDto;
  temperaturePolicy: ModelReasoningCapabilityDtoTemperaturePolicy;
  topPPolicy: ModelReasoningCapabilityDtoTopPPolicy;
}